Change the fallback to get the user object
The line `$vendor_avatar = get_avatar( $the_vendor->ID, 100 );` and others expect $the_vendor to be an object, not an integer. Currently does a fatal error if the if ( !$vendor ) fall back is hit, since get_avatar is expecting $the_vendor to be an object.

On a seperate note, why are you doing `$is_vendor_profile = '';` in the conditional? Its always overriden by the declaration a few lines below. I think `$is_vendor_profile = 'is-vendor-profile';` should be above that fallback if (!vendor)